⚙️ config(server): register update story route
