+++ apps/server/rest-client/space/stars.rest
@@ -10,3 +10,5 @@ Authorization: Bearer {{JWT}}
 GET {{URL}}/id/{{STAR_ID}}
 Authorization: Bearer {{JWT}}
 
+
+
